Move command logic tests into their own test files
